Move vioserialaddrs from QEMU's private data to domainDef
More hypervisors may implement the possibility of specifying device
addresses, so it makes sense to make QEMU-specific code more generic
to be prepared for it.

The end goal is to move all of the address types outside of QEMU's
private data and make the functions from qemu_domain_address.c
hypervisor-agnostic.