Move virtio *Free() functions from domain_addr to domain_conf
More hypervisors may implement the possibility of specifying device
addresses, so it makes sense to make QEMU-specific code more generic
to be prepared for it.

The first step is to move data from obj->privateData to domainDef.
Two functions and a few struct definitions have to be in domain_addr
because virDomainDefFree() has to execute them when freeing
the domainDef structure.