Move functions from qemu_domain_address to domain_addr
These functions don't use any QEMU's specific data, so they can
be moved to a file which is shared between all of the hypervisors.
Their names are changed from qemu* to vir*.

More hypervisors will probably implement the possibility
of explicitly stating the address of devices, so it makes
sense to be prepared for it by making address assignment generic.

After moving the rest of the functions, I will restore the
"static" keyword wherever applicable.