Change parameters to qemuDomainAttachDeviceLive
We want to be able to pass a NULL instead of the connection
and use this function in tests. To achieve this, the virConnectPtr
is passed instead of virDomainPtr, and the driver is a new separate
parameter.

(cherry picked from commit 93fe51c67f461e95ed8d4d16cb5efb49ca868a3b)