qemuhotplugtest: handle xml files more precisely
In qemuhotplugtest.c, the way xml files with domain and device
definitions were handled was very unintuitive. Because of this,
creating new tests was troublesome. Sometimes, you were forced
to create duplicates of files.

I modified the tests to explicitly state the three xml files:
the domain xml, the device xml that is attached/detached/updated,
and the expected xml. Thanks to this, a few duplicates are deleted.
Also, a few xml files that weren't used at all are deleted.