Warn when assigning new pci addresses during recalculation
With a parameter, we can show whether we want
qemuDomainValidateDevicePCISlotsChipsets to assign new addresses or
simply verify the existing ones and reserve slots for potential future
devices. If the parameter is false and some device is missing an address,
a warning will be displayed.
The PCI address set will be recalculated instead of being cached, and
during recalculation all the addresses should have already been
assigned.